requirements.txt
schwabdev
websockets
requests
tk
nicegui
nicegui-highcharts
pywhatkit
pytest_mock
lightweight-charts
scipy
python-dotenv
pandas
anthropic
openai
google-genai